score/requirements.txt
pillow
pytesseract
PyPDF2
openpyxl
